# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Satta Matka Gaming Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/send-otp:
    post:
      operationId: sendOtp
      tags: [auth]
      summary: Send OTP to phone
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendOtpInput"
      responses:
        "200":
          description: OTP sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OtpResponse"
  /auth/verify-otp:
    post:
      operationId: verifyOtp
      tags: [auth]
      summary: Verify OTP
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyOtpInput"
      responses:
        "200":
          description: OTP verified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OtpVerifyResponse"
  /auth/admin-login:
    post:
      operationId: adminLogin
      tags: [auth]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminLoginInput"
      responses:
        "200":
          description: Admin logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"

  # Users
...
  /admin/dashboard:
    get:
      operationId: getAdminDashboard
      tags: [admin]
      summary: Dashboard stats
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Dashboard
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminDashboard"
  /admin/users:
    get:
      operationId: getAdminUsers
      tags: [admin]
      summary: List all users
      security:
        - bearerAuth: []
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Users list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUserListResponse"
  /admin/users/{id}:
    get:
      operationId: getAdminUser
      tags: [admin]
      summary: Get user details (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUserDetail"
    patch:
      operationId: updateAdminUser
      tags: [admin]
      summary: Update user (admin) - block/unblock, change phone
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUserUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /admin/users/{id}/bets:
    get:
      operationId: getAdminUserBets
      tags: [admin]
      summary: Get user bets (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Bets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Bet"
  /admin/bets:
    get:
      operationId: getAdminBets
      tags: [admin]
      summary: All bets (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: marketId
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: date
          in: query
          schema:
            type: string
      responses:
        "200":
          description: All bets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Bet"

  # Settings
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    RegisterInput:
      type: object
      required: [phone, name, password]
      properties:
        phone:
          type: string
        name:
          type: string
        password:
          type: string
        referralCode:
          type: ["string", "null"]
    LoginInput:
      type: object
      required: [phone, password]
      properties:
        phone:
          type: string
        password:
          type: string
    AdminLoginInput:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
    SendOtpInput:
      type: object
      required: [phone]
      properties:
        phone:
          type: string
    VerifyOtpInput:
      type: object
      required: [phone, otp]
      properties:
        phone:
          type: string
        otp:
          type: string
    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
    OtpResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
    OtpVerifyResponse:
      type: object
      required: [verified]
      properties:
        verified:
          type: boolean

    # User
    User:
      type: object
      required: [id, phone, name, walletBalance, isBlocked, role, createdAt]
      properties:
        id:
          type: string
        phone:
          type: string
        name:
          type: string
        walletBalance:
          type: number
        isBlocked:
          type: boolean
        role:
          type: string
        referralCode:
          type: ["string", "null"]
        createdAt:
          type: string
...
    AdminUserUpdate:
      type: object
      properties:
        isBlocked:
          type: boolean
        phone:
          type: string
        name:
          type: string
    AdminUserDetail:
      type: object
      required: [id, phone, name, walletBalance, isBlocked, role, createdAt, totalBets, totalDeposit, totalWithdraw]
      properties:
        id:
          type: string
        phone:
          type: string
        name:
          type: string
        walletBalance:
          type: number
        isBlocked:
          type: boolean
        role:
          type: string
        referralCode:
          type: ["string", "null"]
        createdAt:
          type: string
        totalBets:
          type: integer
        totalDeposit:
          type: number
        totalWithdraw:
          type: number
        bankAccount:
          type: ["string", "null"]
        ifscCode:
          type: ["string", "null"]
        upiId:
          type: ["string", "null"]
        accountName:
          type: ["string", "null"]
        bankName:
          type: ["string", "null"]
    AdminUserListResponse:
      type: object
      required: [users, total, page, limit]
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/User"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer

    # Market
...
    Bet:
      type: object
      required: [id, userId, marketId, betType, number, amount, session, status, createdAt]
      properties:
        id:
          type: string
        userId:
          type: string
        userName:
          type: ["string", "null"]
        userPhone:
          type: ["string", "null"]
        marketId:
          type: string
        marketName:
          type: ["string", "null"]
        betType:
          type: string
        number:
          type: string
        amount:
          type: number
        session:
          type: string
        status:
          type: string
        winAmount:
          type: ["number", "null"]
        createdAt:
          type: string
...
    DepositRequest:
      type: object
      required: [id, userId, amount, status, createdAt]
      properties:
        id:
          type: string
        userId:
          type: string
        userName:
          type: ["string", "null"]
        userPhone:
          type: ["string", "null"]
        amount:
          type: number
        utrNumber:
          type: ["string", "null"]
        screenshotUrl:
          type: ["string", "null"]
        status:
          type: string
        adminNote:
          type: ["string", "null"]
        createdAt:
          type: string
...
    AdminDashboard:
      type: object
      required: [totalUsers, totalBetsToday, totalPayoutToday, totalDepositToday, totalWithdrawToday, pendingDeposits, pendingWithdraws, activeMarkets]
      properties:
        totalUsers:
          type: integer
        totalBetsToday:
          type: integer
        totalPayoutToday:
          type: number
        totalDepositToday:
          type: number
        totalWithdrawToday:
          type: number
        pendingDeposits:
          type: integer
        pendingWithdraws:
          type: integer
        activeMarkets:
          type: integer
        recentBets:
          type: array
          items:
            $ref: "#/components/schemas/Bet"
        recentDeposits:
          type: array
          items:
            $ref: "#/components/schemas/DepositRequest"

    # Settings
...